substack-acquisition/gcs-to-bigquery/requirements.txt
requests==2.31.0
google-cloud-bigquery
google-cloud-secret-manager==2.20.0
